test: split TabService specs by behavior
