+++ tests/unit/background/services/TabService.test.js
@@ -2,191 +2,25 @@
  * @jest-environment jsdom
  * @jest-environment-options {"url": "https://example.com/"}
  *
- * TabService 單元測試
+ * TabService constructor tests
  */
 
 import { jest } from '@jest/globals';
-import { buildHighlight, buildPageRecord } from './serviceTestSupport.js';
+import {
+  TabService,
+  Logger,
+  createTabService,
+  mockInjectionService,
+  mockLogger,
+  resetTabServiceTestState,
+} from './tabServiceTestHarness.js';
 
-const mockLoggerModule = {
-  log: jest.fn(),
-  start: jest.fn(),
-  ready: jest.fn(),
-  info: jest.fn(),
-  warn: jest.fn(),
-  error: jest.fn(),
-  debug: jest.fn(),
-  success: jest.fn(),
-};
-
-const mockCoreConfig = {
-  TAB_SERVICE: {
-    LOADING_TIMEOUT_MS: 1000,
-    STATUS_UPDATE_DELAY_MS: 100,
-    PRELOADER_PING_TIMEOUT_MS: 500,
-  },
-  HANDLER_CONSTANTS: {
-    PAGE_STATUS_CACHE_TTL: 60_000,
-  },
-  RESTRICTED_PROTOCOLS: [
-    'chrome:',
-    'edge:',
-    'about:',
-    'data:',
-    'chrome-extension:',
-    'view-source:',
-    'file:',
-  ],
-};
-
-const mockContentConfig = {
-  URL_NORMALIZATION: {
-    TRACKING_PARAMS: ['utm_source'],
-  },
-};
-
-const mockUrlUtils = {
-  resolveStorageUrl: jest.fn(url => url),
-  buildStableUrlFromNextData: jest.fn(),
-  hasSameOrigin: jest.fn(),
-  normalizeUrl: jest.fn(url => url),
-  isSafeStableUrl: jest.fn(url => {
-    try {
-      const parsedUrl = new URL(url);
-      return parsedUrl.protocol === 'http:' || parsedUrl.protocol === 'https:';
-    } catch {
-      return false;
-    }
-  }),
-  // isRootUrl 預設回傳 false（非根 URL），避免防護邏輯意外攔截正常 URL
-  isRootUrl: jest.fn(() => false),
-};
-
-jest.mock('../../../../scripts/utils/Logger.js', () => ({
-  __esModule: true,
-  default: mockLoggerModule,
-  ...mockLoggerModule,
-}));
-
-jest.mock('../../../../scripts/config/shared/core.js', () => mockCoreConfig);
-
-jest.mock('../../../../scripts/config/shared/content.js', () => mockContentConfig);
-
-jest.mock('../../../../scripts/utils/urlUtils.js', () => mockUrlUtils);
-
-if (typeof jest.unstable_mockModule === 'function') {
-  jest.unstable_mockModule('../../../../scripts/utils/Logger.js', () => ({
-    default: mockLoggerModule,
-    ...mockLoggerModule,
-  }));
-  jest.unstable_mockModule('../../../../scripts/config/shared/core.js', () => mockCoreConfig);
-  jest.unstable_mockModule('../../../../scripts/config/shared/content.js', () => mockContentConfig);
-  jest.unstable_mockModule('../../../../scripts/utils/urlUtils.js', () => mockUrlUtils);
-}
-
-let TabService;
-let _migrationScript;
-let URL_ALIAS_PREFIX;
-let PAGE_PREFIX;
-let HIGHLIGHTS_PREFIX;
-let sanitizeUrlForLogging;
-let Logger;
-let urlUtils;
-
-beforeAll(async () => {
-  ({ TabService, _migrationScript } =
-    await import('../../../../scripts/background/services/TabService.js'));
-  ({ URL_ALIAS_PREFIX, PAGE_PREFIX, HIGHLIGHTS_PREFIX } =
-    await import('../../../../scripts/config/shared/storage.js'));
-  ({ sanitizeUrlForLogging } = await import('../../../../scripts/utils/LogSanitizer.js'));
-  ({ default: Logger } = await import('../../../../scripts/utils/Logger.js'));
-  urlUtils = await import('../../../../scripts/utils/urlUtils.js');
-});
-
-// Mock chrome API
-globalThis.chrome = {
-  action: {
-    setBadgeText: jest.fn(),
-    setBadgeBackgroundColor: jest.fn(),
-  },
-  storage: {
-    local: {
-      get: jest.fn(),
-      set: jest.fn(),
-    },
-    sync: {
-      get: jest.fn(),
-    },
-  },
-  tabs: {
-    onUpdated: {
-      addListener: jest.fn(),
-      removeListener: jest.fn(),
-    },
-    onActivated: {
-      addListener: jest.fn(),
-    },
-    onRemoved: {
-      addListener: jest.fn(),
-      removeListener: jest.fn(),
-    },
-    get: jest.fn(),
-    sendMessage: jest.fn().mockReturnValue(Promise.resolve()),
-    query: jest.fn().mockResolvedValue([]),
-    create: jest.fn().mockResolvedValue({}),
-    remove: jest.fn().mockResolvedValue(),
-  },
-};
-
-// Mock Logger
-const mockLogger = {
-  log: jest.fn(),
-  start: jest.fn(),
-  ready: jest.fn(),
-  warn: jest.fn(),
-  error: jest.fn(),
-  debug: jest.fn(),
-  info: jest.fn(),
-  success: jest.fn(),
-};
-
-// Mock InjectionService
-const mockInjectionService = {
-  ensureBundleInjected: jest.fn().mockResolvedValue(true),
-  injectWithResponse: jest.fn().mockResolvedValue({ migrated: false }),
-  injectHighlightRestore: jest.fn().mockResolvedValue(),
-};
-
-const createTabService = (overrides = {}) =>
-  new TabService({
-    logger: mockLogger,
-    injectionService: mockInjectionService,
-    normalizeUrl: url => url,
-    getSavedPageData: jest.fn().mockResolvedValue(null),
-    isRestrictedUrl: url => url.includes('chrome://'),
-    isRecoverableError: err => {
-      const msg = typeof err === 'string' ? err : err?.message || '';
-      return msg.includes('Cannot access');
-    },
-    checkPageExists: jest.fn().mockResolvedValue(true),
-    getApiKey: jest.fn().mockResolvedValue('test-api-key'),
-    clearPageState: jest.fn().mockResolvedValue(),
-    clearNotionState: jest.fn().mockResolvedValue(),
-    clearNotionStateWithRetry: jest.fn().mockResolvedValue({ cleared: true, attempts: 1 }),
-    setSavedPageData: jest.fn().mockResolvedValue(),
-    ...overrides,
-  });
-
-describe('TabService', () => {
+describe('TabService constructor', () => {
   let service = null;
 
   beforeEach(() => {
+    resetTabServiceTestState();
     service = createTabService();
-
-    // 初始化全局 chrome.runtime
-    chrome.runtime = { lastError: null };
-
-    jest.clearAllMocks();
   });
 
   describe('constructor', () => {
@@ -264,1493 +98,6 @@ describe('TabService', () => {
         attempts: 1,
         recovered: false,
       });
-    });
-  });
-
-  describe('updateTabStatus', () => {
-    it('should skip invalid URLs', async () => {
-      await service.updateTabStatus(1, null);
-      await service.updateTabStatus(1, '');
-      await service.updateTabStatus(1, 'ftp://example.com');
-
-      expect(chrome.action.setBadgeText).not.toHaveBeenCalled();
-    });
-
-    it('should skip restricted URLs', async () => {
-      await service.updateTabStatus(1, 'chrome://extensions');
-
-      expect(chrome.action.setBadgeText).not.toHaveBeenCalled();
-    });
-
-    it('should call restricted URL dependency with the service context', async () => {
-      service.isRestrictedUrl = jest.fn(function () {
-        return this === service;
-      });
-
-      await service.updateTabStatus(1, 'https://example.com');
-
-      expect(service.isRestrictedUrl).toHaveBeenCalledWith('https://example.com');
-      expect(service.getSavedPageData).not.toHaveBeenCalled();
-      expect(chrome.action.setBadgeText).not.toHaveBeenCalled();
-    });
-
-    it('should set badge for saved pages', async () => {
-      service.getSavedPageData = jest.fn().mockResolvedValue({
-        pageId: '123',
-        notionPageId: 'notion-123',
-        lastVerifiedAt: Date.now(), // 確保在 TTL 內
-      });
-      chrome.storage.local.get.mockImplementation(_keys => Promise.resolve({}));
-
-      await service.updateTabStatus(1, 'https://example.com');
-
-      expect(chrome.action.setBadgeText).toHaveBeenCalledWith({ text: '✓', tabId: 1 });
-      expect(chrome.action.setBadgeBackgroundColor).toHaveBeenCalledWith({
-        color: '#48bb78',
-        tabId: 1,
-      });
-    });
-
-    it('should clear badge for unsaved pages', async () => {
-      service.getSavedPageData = jest.fn().mockResolvedValue(null);
-      chrome.storage.local.get.mockImplementation(_keys => Promise.resolve({}));
-
-      await service.updateTabStatus(1, 'https://example.com');
-
-      expect(chrome.action.setBadgeText).toHaveBeenCalledWith({ text: '', tabId: 1 });
-    });
-
-    it('should inject bundle for auto-restore when highlights exist', async () => {
-      // Mock highlights 存在
-      service.getSavedPageData = jest.fn().mockResolvedValue(null);
-      chrome.storage.local.get.mockImplementation(_keys =>
-        Promise.resolve({ 'highlights_https://example.com': [{ id: '1' }] })
-      );
-
-      // Mock tab get 返回 complete 狀態
-      chrome.tabs.get.mockImplementation(_tabId =>
-        Promise.resolve({ id: 1, status: 'complete', url: 'https://example.com' })
-      );
-
-      await service.updateTabStatus(1, 'https://example.com');
-
-      // 驗證使用 ensureBundleInjected 而非 injectHighlighter
-      expect(mockInjectionService.ensureBundleInjected).toHaveBeenCalledWith(1);
-    });
-
-    it('should normalize originalUrl when creating fallback url_alias key', async () => {
-      service.resolveTabUrl = jest.fn().mockResolvedValue({
-        stableUrl: 'https://example.com/stable',
-        originalUrl: 'https://example.com/original/?utm_source=fb#frag',
-        hasStableUrl: true,
-      });
-      service._verifyAndUpdateStatus = jest.fn().mockResolvedValue();
-      service._getHighlightsFromStorage = jest
-        .fn()
-        .mockResolvedValueOnce(null)
-        .mockResolvedValueOnce([{ id: '1' }]);
-      service._waitForTabCompilation = jest
-        .fn()
-        .mockResolvedValue({ id: 1, status: 'complete', url: 'https://example.com/stable' });
-      service._sendStableUrl = jest.fn();
-      service.normalizeUrl = jest.fn(url => url.replace('/?utm_source=fb#frag', ''));
-
-      // Phase 4 (tighten)：alias 寫入需 stableUrl 已有有效 evidence；seed page_<stable> 含 highlights
-      chrome.storage.local.get.mockResolvedValue({
-        'page_https://example.com/stable': {
-          notion: null,
-          highlights: [{ id: 'evidence-h', text: 'evidence', color: 'yellow' }],
-          metadata: { lastUpdated: 1 },
-        },
-      });
-      chrome.storage.local.set.mockResolvedValue(undefined);
-
-      await service._updateTabStatusInternal(1, 'https://example.com/original/?utm_source=fb#frag');
-
-      expect(chrome.storage.local.set).toHaveBeenCalledWith(
-        expect.objectContaining({
-          [`${URL_ALIAS_PREFIX}https://example.com/original`]: 'https://example.com/stable',
-        })
-      );
-    });
-
-    it('Phase 4 (tighten)：hasStableUrl=true 但 stableUrl 無 evidence 時，MUST NOT 寫入 url_alias', async () => {
-      // 情境：Preloader 解析出 stableUrl，但 page_<stable> 與 highlights_<stable> 都無資料
-      // 預期：tighten 模式下 alias 寫入應被跳過，避免 canonical 漂移
-      service.resolveTabUrl = jest.fn().mockResolvedValue({
-        stableUrl: 'https://example.com/?p=2928',
-        originalUrl: 'https://example.com/long-path',
-        hasStableUrl: true,
-      });
-      service._verifyAndUpdateStatus = jest.fn().mockResolvedValue();
-      service._getHighlightsFromStorage = jest.fn().mockResolvedValue(null);
-      jest.spyOn(service, 'migrateLegacyHighlights').mockResolvedValue();
-
-      // 無 evidence：所有 key 都返回空
-      chrome.storage.local.get.mockResolvedValue({});
-      chrome.storage.local.set.mockResolvedValue(undefined);
-
-      await service._updateTabStatusInternal(1, 'https://example.com/long-path');
-
-      // alias key MUST NOT 被寫入
-      const aliasSets = chrome.storage.local.set.mock.calls.filter(call =>
-        Object.keys(call[0]).some(k => k.startsWith(URL_ALIAS_PREFIX))
-      );
-      expect(aliasSets).toHaveLength(0);
-    });
-
-    it('Phase 4 (tighten)：hasStableUrl=true 且 stableUrl 已有 page_* evidence 時，建立 url_alias', async () => {
-      // 情境：stableUrl 對應的 page_* 已有 highlights → 滿足 tighten 條件
-      service.resolveTabUrl = jest.fn().mockResolvedValue({
-        stableUrl: 'https://example.com/?p=2928',
-        originalUrl: 'https://example.com/long-path',
-        hasStableUrl: true,
-      });
-      service._verifyAndUpdateStatus = jest.fn().mockResolvedValue();
-      service._getHighlightsFromStorage = jest.fn().mockResolvedValue(null);
-      jest.spyOn(service, 'migrateLegacyHighlights').mockResolvedValue();
-
-      chrome.storage.local.get.mockResolvedValue({
-        'page_https://example.com/?p=2928': {
-          notion: null,
-          highlights: [{ id: 'h', text: 'evidence', color: 'yellow' }],
-          metadata: { lastUpdated: 1 },
-        },
-      });
-      chrome.storage.local.set.mockResolvedValue(undefined);
-
-      await service._updateTabStatusInternal(1, 'https://example.com/long-path');
-
-      // alias 應該在 evidence 滿足時建立
-      expect(chrome.storage.local.set).toHaveBeenCalledWith({
-        [`${URL_ALIAS_PREFIX}https://example.com/long-path`]: 'https://example.com/?p=2928',
-      });
-    });
-
-    it('應等待 url_alias 寫入完成後才繼續更新狀態', async () => {
-      service.resolveTabUrl = jest.fn().mockResolvedValue({
-        stableUrl: 'https://example.com/?p=2928',
-        originalUrl: 'https://example.com/long-path',
-        hasStableUrl: true,
-      });
-      service._verifyAndUpdateStatus = jest.fn().mockResolvedValue();
-      service._getHighlightsFromStorage = jest.fn().mockResolvedValue(null);
-      jest.spyOn(service, 'migrateLegacyHighlights').mockResolvedValue();
-
-      // Phase 4 follow-up（2026-05-03 plan §4）：顯式 mock chrome.storage.local.get,
-      // 提供 stableUrl 的 page_* evidence,讓 _persistUrlAliasIfNeeded 真的走到 set 寫入路徑。
-      // 不再依賴前一 test 的 mockResolvedValue 殘留（jest.clearAllMocks 不清 implementation,
-      // 順序變動會讓本 test 偷偷走到 throw catch 路徑而非真的等 set 完成）。
-      chrome.storage.local.get.mockResolvedValue({
-        'page_https://example.com/?p=2928': {
-          notion: null,
-          highlights: [{ id: 'evidence-h', text: 'evidence', color: 'yellow' }],
-          metadata: { lastUpdated: 1 },
-        },
-      });
-
-      let resolveStorageSet;
-      chrome.storage.local.set.mockReturnValue(
-        new Promise(resolve => {
-          resolveStorageSet = resolve;
-        })
-      );
-
-      const updatePromise = service._updateTabStatusInternal(1, 'https://example.com/long-path');
-      await Promise.resolve();
-
-      expect(service._verifyAndUpdateStatus).not.toHaveBeenCalled();
-
-      resolveStorageSet();
-      await updatePromise;
-
-      expect(service._verifyAndUpdateStatus).toHaveBeenCalledWith(
-        1,
-        'https://example.com/?p=2928',
-        'https://example.com/long-path'
-      );
-    });
-
-    it('url_alias 寫入失敗時應記錄警告並繼續更新狀態', async () => {
-      service.resolveTabUrl = jest.fn().mockResolvedValue({
-        stableUrl: 'https://example.com/?p=2928',
-        originalUrl: 'https://example.com/long-path',
-        hasStableUrl: true,
-      });
-      service._verifyAndUpdateStatus = jest.fn().mockResolvedValue();
-      service._getHighlightsFromStorage = jest.fn().mockResolvedValue(null);
-      jest.spyOn(service, 'migrateLegacyHighlights').mockResolvedValue();
-
-      // Phase 4 (tighten)：先 seed evidence 才會走到 alias 寫入路徑，從而觸發 set 失敗
-      chrome.storage.local.get.mockResolvedValue({
-        'page_https://example.com/?p=2928': {
-          notion: null,
-          highlights: [{ id: 'h', text: 'evidence', color: 'yellow' }],
-          metadata: { lastUpdated: 1 },
-        },
-      });
-      chrome.storage.local.set.mockRejectedValue(new Error('alias write failed'));
-
-      await service._updateTabStatusInternal(1, 'https://example.com/long-path');
-
-      expect(service._verifyAndUpdateStatus).toHaveBeenCalledWith(
-        1,
-        'https://example.com/?p=2928',
-        'https://example.com/long-path'
-      );
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        '[TabService] url_alias 寫入或檢查失敗，將繼續後續狀態更新',
-        expect.objectContaining({
-          action: 'updateTabStatus',
-          originalUrl: expect.any(String),
-          stableUrl: expect.any(String),
-          error: expect.any(Error),
-        })
-      );
-    });
-
-    it('hasStableUrl=false 時不應建立 url_alias', async () => {
-      // 情境：Preloader 超時，stableUrl = originalUrl
-      service.resolveTabUrl = jest.fn().mockResolvedValue({
-        stableUrl: 'https://example.com/long-path',
-        originalUrl: 'https://example.com/long-path',
-        hasStableUrl: false,
-      });
-      service._verifyAndUpdateStatus = jest.fn().mockResolvedValue();
-      service._getHighlightsFromStorage = jest.fn().mockResolvedValue(null);
-      jest.spyOn(service, 'migrateLegacyHighlights').mockResolvedValue();
-      chrome.storage.local.set.mockResolvedValue(undefined);
-
-      await service._updateTabStatusInternal(1, 'https://example.com/long-path');
-
-      // 不應有任何 alias 建立調用（storage.local.set 未因 alias 而被呼叫）
-      const aliasCalls = chrome.storage.local.set.mock.calls.filter(([arg]) =>
-        Object.keys(arg).some(k => k.startsWith(URL_ALIAS_PREFIX))
-      );
-      expect(aliasCalls).toHaveLength(0);
-    });
-
-    it('stableUrl 查無 highlights 時應 fallback 到 originalUrl 並成功回傳 highlights', async () => {
-      // 驗證 fallback 資料命中時確實回傳 highlights（功能正確性）
-      const highlights = [{ id: 'fallback-highlight' }];
-
-      service._getHighlightsFromStorage = jest
-        .fn()
-        .mockResolvedValueOnce(null)
-        .mockResolvedValueOnce(highlights);
-
-      const result = await service._getHighlightsWithFallback(
-        'https://example.com/stable',
-        true,
-        'https://example.com/original'
-      );
-
-      expect(result).toBe(highlights);
-      expect(service._getHighlightsFromStorage).toHaveBeenNthCalledWith(
-        1,
-        'https://example.com/stable'
-      );
-      expect(service._getHighlightsFromStorage).toHaveBeenNthCalledWith(
-        2,
-        'https://example.com/original'
-      );
-    });
-
-    it('應以脫敏後的 highlight storage key 記錄除錯資訊', async () => {
-      const rawUrl = 'https://example.com/article?token=secret123&utm_source=ads#frag';
-      const rawStorageKey = `page_${rawUrl}`;
-      chrome.storage.local.get.mockResolvedValue({
-        [rawStorageKey]: { highlights: [{ id: 'highlight-1' }] },
-      });
-
-      await service._getHighlightsFromStorage(rawUrl);
-
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        `[TabService] Checking highlights for page_${sanitizeUrlForLogging(rawUrl)}:`,
-        {
-          found: true,
-          count: 1,
-        }
-      );
-      expect(mockLogger.debug).not.toHaveBeenCalledWith(
-        expect.stringContaining(rawStorageKey),
-        expect.anything()
-      );
-    });
-
-    it('[REGRESSION] url_alias 指向 stableUrl 時，應能命中 page_<stableUrl>', async () => {
-      const originalUrl = 'https://example.com/posts/hello';
-      const stableUrl = 'https://example.com/?p=123';
-      const stablePageKey = `${PAGE_PREFIX}${stableUrl}`;
-      const legacyKey = `${HIGHLIGHTS_PREFIX}${originalUrl}`;
-      const highlights = [{ id: 'stable-highlight' }];
-
-      chrome.storage.local.get.mockImplementation(async keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-        const result = {};
-        if (keyList.includes(`${URL_ALIAS_PREFIX}${originalUrl}`)) {
-          result[`${URL_ALIAS_PREFIX}${originalUrl}`] = stableUrl;
-        }
-        if (keyList.includes(stablePageKey)) {
-          result[stablePageKey] = { highlights };
-        }
-        return result;
-      });
-
-      const result = await service._getHighlightsFromStorage(originalUrl);
-
-      expect(result).toEqual(highlights);
-      expect(chrome.storage.local.get).toHaveBeenNthCalledWith(
-        1,
-        expect.arrayContaining([
-          `${URL_ALIAS_PREFIX}${originalUrl}`,
-          `${PAGE_PREFIX}${originalUrl}`,
-          legacyKey,
-        ])
-      );
-      expect(chrome.storage.local.get).toHaveBeenNthCalledWith(
-        2,
-        expect.arrayContaining([stablePageKey, `${HIGHLIGHTS_PREFIX}${stableUrl}`])
-      );
-    });
-
-    // ─── Step 0.2 Regression Test ────────────────────────────────────────────
-    // 驗證：fallback 命中 originalUrl 時，不應無條件補寫 url_alias。
-    // 修復前此測試應 FAIL（因為現有程式碼會無條件補寫 alias）。
-    it('[REGRESSION] fallback 命中 page_<originalUrl> 時，不應立即補寫 url_alias', async () => {
-      const highlights = [{ id: 'regression-no-alias' }];
-
-      service._getHighlightsFromStorage = jest
-        .fn()
-        .mockResolvedValueOnce(null) // stableUrl miss
-        .mockResolvedValueOnce(highlights); // originalUrl hit
-
-      await service._getHighlightsWithFallback(
-        'https://example.com/stable',
-        true,
-        'https://example.com/original'
-      );
-
-      // 修復後：不應在 fallback 命中後補寫 alias
-      const aliasCalls = chrome.storage.local.set.mock.calls.filter(([arg]) =>
-        Object.keys(arg).some(k => k.startsWith(URL_ALIAS_PREFIX))
-      );
-      expect(aliasCalls).toHaveLength(0);
-    });
-
-    it('refresh 後解析出 stableUrl 且僅有 highlights 時，應保持未保存 badge 並注入 bundle', async () => {
-      const originalUrl = 'https://example.com/articles/slug';
-      const stableUrl = 'https://example.com/?p=2928';
-
-      service.getSavedPageData = jest.fn().mockResolvedValue(null);
-      service.resolveTabUrl = jest
-        .fn()
-        .mockResolvedValueOnce({
-          stableUrl: originalUrl,
-          originalUrl,
-          hasStableUrl: false,
-        })
-        .mockResolvedValueOnce({
-          stableUrl,
-          originalUrl,
-          hasStableUrl: true,
-        });
-
-      chrome.storage.local.get.mockImplementation(keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-
-        if (keyList.includes(`page_${stableUrl}`)) {
-          return Promise.resolve({
-            [`page_${stableUrl}`]: buildPageRecord({
-              notion: null,
-              highlights: [buildHighlight()],
-            }),
-          });
-        }
-
-        return Promise.resolve({});
-      });
-
-      chrome.tabs.get.mockResolvedValue({ id: 1, status: 'complete', url: stableUrl });
-
-      await service.updateTabStatus(1, originalUrl);
-      await service.updateTabStatus(1, originalUrl);
-
-      expect(chrome.action.setBadgeText).toHaveBeenLastCalledWith({ text: '', tabId: 1 });
-      expect(mockInjectionService.ensureBundleInjected).toHaveBeenCalledWith(1);
-      expect(chrome.storage.local.set).toHaveBeenCalledWith({
-        [`${URL_ALIAS_PREFIX}${originalUrl}`]: stableUrl,
-      });
-    });
-
-    it('refresh 後 stableUrl 命中 notion:null 的 page_* 時，不應回彈為已保存', async () => {
-      const originalUrl = 'https://example.com/posts/long-title';
-      const stableUrl = 'https://example.com/?p=123';
-
-      service.resolveTabUrl = jest
-        .fn()
-        .mockResolvedValueOnce({
-          stableUrl: originalUrl,
-          originalUrl,
-          hasStableUrl: false,
-        })
-        .mockResolvedValueOnce({
-          stableUrl,
-          originalUrl,
-          hasStableUrl: true,
-        });
-      service.getSavedPageData = jest.fn().mockImplementation(url => {
-        if (url === originalUrl || url === stableUrl) {
-          return Promise.resolve(null);
-        }
-        return Promise.resolve(null);
-      });
-
-      chrome.storage.local.get.mockImplementation(keys => {
-        const keyList = Array.isArray(keys) ? keys : [keys];
-
-        if (keyList.includes(`page_${stableUrl}`)) {
-          return Promise.resolve({
-            [`page_${stableUrl}`]: buildPageRecord({
-              notion: null,
-              highlights: [],
-            }),
-          });
-        }
-
-        return Promise.resolve({});
-      });
-
-      chrome.tabs.get.mockResolvedValue({ id: 1, status: 'complete', url: stableUrl });
-
-      await service.updateTabStatus(1, originalUrl);
-      await service.updateTabStatus(1, originalUrl);
-
-      expect(chrome.action.setBadgeText).toHaveBeenLastCalledWith({ text: '', tabId: 1 });
-      // v2.68.0+: Floating Rail 預設啟用，無歷史標註頁面也應注入 bundle 讓 rail 自動載入
-      expect(mockInjectionService.ensureBundleInjected).toHaveBeenCalledWith(1);
-    });
-
-    it('should call migrateLegacyHighlights when no highlights exist', async () => {
-      service.getSavedPageData = jest.fn().mockResolvedValue(null);
-      chrome.storage.local.get.mockImplementation(_keys => Promise.resolve({}));
-
-      const migrateSpy = jest.spyOn(service, 'migrateLegacyHighlights').mockResolvedValue();
-
-      await service.updateTabStatus(1, 'https://example.com');
-
-      expect(migrateSpy).toHaveBeenCalledWith(
-        1,
-        'https://example.com',
-        'highlights_https://example.com'
-      );
-    });
-
-    it('[REGRESSION] Floating Rail 預設啟用時，無歷史標註的頁面也應注入 content bundle', async () => {
-      // Bug：v2.68.0 加入 Floating Rail 後，TabService 在 `if (!highlights)` 分支
-      // 直接 return，導致全新網頁從未注入 content.bundle.js，rail 永遠不出現。
-      // 修復：在無 highlights 分支內仍判斷 floatingRailEnabled，預設啟用即注入 bundle。
-      service.getSavedPageData = jest.fn().mockResolvedValue(null);
-      chrome.storage.local.get.mockImplementation(_keys => Promise.resolve({}));
-      chrome.storage.sync.get = jest.fn().mockResolvedValue({}); // 未設定 = 預設啟用
-      chrome.tabs.get.mockResolvedValue({
-        id: 1,
-        status: 'complete',
-        url: 'https://example.com',
-      });
-      jest.spyOn(service, 'migrateLegacyHighlights').mockResolvedValue();
-
-      await service.updateTabStatus(1, 'https://example.com');
-
-      expect(mockInjectionService.ensureBundleInjected).toHaveBeenCalledWith(1);
-    });
-
-    it('[REGRESSION] floatingRailEnabled=false 時，無歷史標註頁面不應注入 bundle', async () => {
-      service.getSavedPageData = jest.fn().mockResolvedValue(null);
-      chrome.storage.local.get.mockImplementation(_keys => Promise.resolve({}));
-      chrome.storage.sync.get = jest.fn().mockResolvedValue({ floatingRailEnabled: false });
-      jest.spyOn(service, 'migrateLegacyHighlights').mockResolvedValue();
-
-      await service.updateTabStatus(1, 'https://example.com');
-
-      expect(mockInjectionService.ensureBundleInjected).not.toHaveBeenCalled();
-    });
-
-    it('should handle ensureBundleInjected rejection gracefully', async () => {
-      // Arrange: 模擬 highlights 存在
-      service.getSavedPageData = jest.fn().mockResolvedValue(null);
-      chrome.storage.local.get.mockImplementation(_keys =>
-        Promise.resolve({ 'highlights_https://example.com': [{ id: '1' }] })
-      );
-
-      chrome.tabs.get.mockImplementation(_tabId => {
-        chrome.runtime = { lastError: null };
-        return Promise.resolve({ id: 1, status: 'complete', url: 'https://example.com' });
-      });
-
-      // Arrange: 模擬注入失敗
-      const injectionError = new Error('Bundle injection failed');
-      mockInjectionService.ensureBundleInjected.mockRejectedValue(injectionError);
-
-      // Act
-      await service.updateTabStatus(1, 'https://example.com');
-
-      // Assert: 錯誤被記錄
-      expect(mockLogger.error).toHaveBeenCalledWith(
-        expect.stringContaining('[TabService] Error updating tab status'),
-        expect.objectContaining({ error: injectionError })
-      );
-
-      // Assert: ensureBundleInjected 被調用
-      expect(mockInjectionService.ensureBundleInjected).toHaveBeenCalledWith(1);
-    });
-
-    it('should handle errors gracefully', async () => {
-      // 確保 getSavedPageData 拋出錯誤
-      service.getSavedPageData = jest.fn().mockRejectedValue(new Error('Storage error'));
-
-      await service.updateTabStatus(1, 'https://example.com');
-
-      expect(mockLogger.error).toHaveBeenCalled();
-    });
-
-    describe('Automatic Verification', () => {
-      it('should verify with Notion when cache is expired', async () => {
-        const expiredData = {
-          notionPageId: 'page-123',
-          lastVerifiedAt: Date.now() - 70_000, // 超過 60s
-        };
-        service.getSavedPageData = jest.fn().mockResolvedValue(expiredData);
-        service.checkPageExists = jest.fn().mockResolvedValue(true);
-
-        await service.updateTabStatus(1, 'https://example.com');
-
-        expect(service.checkPageExists).toHaveBeenCalledWith('page-123', 'test-api-key');
-        expect(service.setSavedPageData).toHaveBeenCalledWith(
-          'https://example.com',
-          expect.objectContaining({ lastVerifiedAt: expect.any(Number) })
-        );
-      });
-
-      it('should clear local state only after two consecutive false checks', async () => {
-        const expiredData = {
-          notionPageId: 'page-123',
-          lastVerifiedAt: Date.now() - 70_000,
-        };
-        service.getSavedPageData = jest.fn().mockResolvedValue(expiredData);
-        service.checkPageExists = jest.fn().mockResolvedValue(false); // 模擬已刪除
-
-        await service.updateTabStatus(1, 'https://example.com');
-        expect(service.clearNotionStateWithRetry).not.toHaveBeenCalled();
-
-        await service.updateTabStatus(1, 'https://example.com');
-        expect(service.clearNotionStateWithRetry).toHaveBeenCalledWith(
-          'https://example.com',
-          expect.objectContaining({ source: 'TabService._handleNotionVerificationResult' })
-        );
-        expect(chrome.action.setBadgeText).toHaveBeenCalledWith({ text: '', tabId: 1 });
-      });
-
-      it('should preserve saved badge and avoid re-arming deletion when cleanup is skipped', async () => {
-        const expiredData = {
-          notionPageId: 'page-123',
-          lastVerifiedAt: Date.now() - 70_000,
-        };
-        service.getSavedPageData = jest.fn().mockResolvedValue(expiredData);
-        service.checkPageExists = jest.fn().mockResolvedValue(false);
-        service.clearNotionStateWithRetry = jest.fn().mockImplementation(async (_url, options) => {
-          if (options?.expectedPageId === 'page-123') {
-            return {
-              cleared: false,
-              skipped: true,
-              reason: 'pageId_mismatch',
-              attempts: 1,
-              recovered: false,
-            };
-          }
-
-          return {
-            cleared: true,
-            attempts: 1,
-            recovered: false,
-          };
-        });
-
-        await service.updateTabStatus(1, 'https://example.com');
-        await service.updateTabStatus(1, 'https://example.com');
-
-        expect(service.clearNotionStateWithRetry).toHaveBeenCalledWith(
-          'https://example.com',
-          expect.objectContaining({
-            source: 'TabService._handleNotionVerificationResult',
-            expectedPageId: 'page-123',
-          })
-        );
-        expect(mockLogger.warn).toHaveBeenCalledWith(
-          '[TabService] 清理已略過：本機 Notion 綁定已變更',
-          expect.objectContaining({
-            action: 'autoSyncLocalState',
-            pageId: 'page',
-            reason: 'pageId_mismatch',
-            result: 'cleanup_skipped',
-          })
-        );
-        expect(service.deletionPendingPages.has('page-123')).toBe(false);
-        expect(chrome.action.setBadgeText).toHaveBeenLastCalledWith({ text: '✓', tabId: 1 });
-      });
-
-      it('should fallback to cached status if verification fails', async () => {
-        const expiredData = {
-          pageId: '123',
-          notionPageId: 'page-123',
-          lastVerifiedAt: Date.now() - 70_000,
-        };
-        service.getSavedPageData = jest.fn().mockResolvedValue(expiredData);
-        service.checkPageExists = jest.fn().mockRejectedValue(new Error('Notion API Error'));
-
-        await service.updateTabStatus(1, 'https://example.com');
-
-        expect(mockLogger.warn).toHaveBeenCalled();
-        expect(chrome.action.setBadgeText).toHaveBeenCalledWith({ text: '✓', tabId: 1 });
-        // 依然顯示勾勾
-      });
-
-      it('getApiKey 返回 null 時不應清除 deletionPendingPages（OAuth 用戶迴歸）', async () => {
-        // 情境：用戶使用 OAuth 模式，TabService.getApiKey 錯誤地返回 null
-        // 預期行為：pending 狀態應被保留，不應被清除
-        const expiredData = {
-          notionPageId: 'page-123',
-          lastVerifiedAt: Date.now() - 70_000,
-        };
-        service.getSavedPageData = jest.fn().mockResolvedValue(expiredData);
-        service.getApiKey = jest.fn().mockResolvedValue(null); // 模擬 OAuth 用戶取不到 key
-
-        // 先模擬 checkPageStatus 已將此頁面標記為 pending
-        service.consumeDeletionConfirmation('page-123', false);
-
-        // 執行背景的 tab 狀態更新（由 onActivated / onUpdated 觸發）
-        await service.updateTabStatus(1, 'https://example.com');
-
-        // 關鍵斷言：getApiKey 返回 null 時，不應重置 deletionPendingPages
-        expect(service.deletionPendingPages.has('page-123')).toBe(true);
-        // badge 應仍顯示已保存
-        expect(chrome.action.setBadgeText).toHaveBeenCalledWith({ text: '✓', tabId: 1 });
-        // 不應呼叫 checkPageExists（因為沒有 apiKey）
-        expect(service.checkPageExists).not.toHaveBeenCalled();
-      });
-
-      it('should skip Verification if notionPageId is missing', async () => {
-        const expiredData = {
-          lastVerifiedAt: Date.now() - 70_000,
-          // 故意缺少 notionPageId
-        };
-        service.getSavedPageData = jest.fn().mockResolvedValue(expiredData);
-
-        await service.updateTabStatus(1, 'https://example.com');
-        expect(mockLogger.debug).toHaveBeenCalledWith(
-          expect.stringContaining('No notionPageId in savedData')
-        );
-        expect(service.checkPageExists).not.toHaveBeenCalled();
-        expect(chrome.action.setBadgeText).toHaveBeenCalledWith({ text: '✓', tabId: 1 });
-      });
-
-      it('should mark as pending on first deletion check failure', async () => {
-        const expiredData = {
-          notionPageId: 'page-123',
-          lastVerifiedAt: Date.now() - 70_000,
-        };
-        service.getSavedPageData = jest.fn().mockResolvedValue(expiredData);
-        service.checkPageExists = jest.fn().mockResolvedValue(false);
-        // 不 mock consumeDeletionConfirmation，讓它真實回傳 { deletionPending: true }
-
-        await service.updateTabStatus(1, 'https://example.com');
-
-        expect(mockLogger.warn).toHaveBeenCalledWith(
-          expect.stringContaining('First deletion check failed'),
-          expect.objectContaining({
-            pageId: 'page',
-            action: 'autoSyncLocalState',
-          })
-        );
-        expect(chrome.action.setBadgeText).toHaveBeenCalledWith({ text: '✓', tabId: 1 });
-      });
-
-      it('should update lastVerifiedAt if page remains true', async () => {
-        const expiredData = {
-          notionPageId: 'page-123',
-          lastVerifiedAt: Date.now() - 70_000,
-        };
-        service.getSavedPageData = jest.fn().mockResolvedValue(expiredData);
-        service.checkPageExists = jest.fn().mockResolvedValue(true);
-
-        await service.updateTabStatus(1, 'https://example.com');
-
-        expect(service.setSavedPageData).toHaveBeenCalledWith(
-          'https://example.com',
-          expect.objectContaining({ lastVerifiedAt: expect.any(Number) })
-        );
-        expect(chrome.action.setBadgeText).toHaveBeenCalledWith({ text: '✓', tabId: 1 });
-      });
-    });
-  });
-
-  describe('remote page missing confirmation API', () => {
-    // 與 TabService 中 DELETION_CONFIRMATION_WINDOW_MS 一致（5 分鐘）
-    const WINDOW_MS = 5 * 60 * 1000;
-
-    afterEach(() => {
-      jest.restoreAllMocks();
-    });
-
-    it('confirmRemotePageMissing should treat expired pending deletion as a new first failure', () => {
-      jest
-        .spyOn(Date, 'now')
-        .mockReturnValueOnce(1000)
-        .mockReturnValueOnce(1000 + WINDOW_MS + 1); // 窗口過期 +1ms
-
-      expect(service.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: false,
-        deletionPending: true,
-      });
-
-      expect(service.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: false,
-        deletionPending: true,
-      });
-    });
-
-    it('resetRemotePageMissingState should clear pending deletion state', () => {
-      expect(service.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: false,
-        deletionPending: true,
-      });
-
-      expect(service.resetRemotePageMissingState('page-1')).toEqual({
-        shouldDelete: false,
-        deletionPending: false,
-      });
-
-      expect(service.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: false,
-        deletionPending: true,
-      });
-    });
-
-    it('confirmRemotePageMissing should delete only on a second false within the confirmation window', () => {
-      jest.spyOn(Date, 'now').mockReturnValue(10_000);
-
-      expect(service.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: false,
-        deletionPending: true,
-      });
-
-      expect(service.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: true,
-        deletionPending: false,
-      });
-    });
-
-    it('should treat pending deletion state as volatile after service re-instantiation', () => {
-      jest.spyOn(Date, 'now').mockReturnValue(30_000);
-
-      expect(service.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: false,
-        deletionPending: true,
-      });
-
-      const restartedService = createTabService();
-
-      expect(restartedService.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: false,
-        deletionPending: true,
-      });
-      expect(restartedService.deletionPendingPages.has('page-1')).toBe(true);
-
-      expect(restartedService.confirmRemotePageMissing('page-1')).toEqual({
-        shouldDelete: true,
-        deletionPending: false,
-      });
-    });
-
-    it('consumeDeletionConfirmation should remain backward compatible', () => {
-      jest.spyOn(Date, 'now').mockReturnValue(20_000);
-
-      expect(service.consumeDeletionConfirmation('page-1', false)).toEqual({
-        shouldDelete: false,
-        deletionPending: true,
-      });
-
-      expect(service.consumeDeletionConfirmation('page-1', null)).toEqual({
-        shouldDelete: false,
-        deletionPending: false,
-      });
-    });
-  });
-
-  describe('setupListeners', () => {
-    it('should register tab update listener', () => {
-      service.setupListeners();
-
-      expect(chrome.tabs.onUpdated.addListener).toHaveBeenCalled();
-    });
-
-    it('should register tab activation listener', () => {
-      service.setupListeners();
-
-      expect(chrome.tabs.onActivated.addListener).toHaveBeenCalled();
-    });
-  });
-
-  describe('migrateLegacyHighlights', () => {
-    beforeEach(() => {
-      chrome.tabs.get.mockResolvedValue({ id: 1, url: 'https://example.com', status: 'complete' });
-    });
-
-    it('should skip if normUrl is missing', async () => {
-      await service.migrateLegacyHighlights(1, null, 'key');
-
-      expect(mockLogger.warn).toHaveBeenCalled();
-      expect(mockInjectionService.injectWithResponse).not.toHaveBeenCalled();
-    });
-
-    it('should skip if storageKey is missing', async () => {
-      await service.migrateLegacyHighlights(1, 'https://example.com', null);
-
-      expect(mockLogger.warn).toHaveBeenCalled();
-    });
-
-    it('should skip non-http URLs', async () => {
-      await service.migrateLegacyHighlights(1, 'ftp://example.com', 'key');
-
-      expect(mockLogger.warn).toHaveBeenCalled();
-    });
-
-    it('should skip if tab is invalid', async () => {
-      chrome.tabs.get.mockRejectedValue(new Error('Tab not found'));
-
-      await service.migrateLegacyHighlights(1, 'https://example.com', 'key');
-
-      expect(mockInjectionService.injectWithResponse).not.toHaveBeenCalled();
-    });
-
-    it('should skip if tab shows error page', async () => {
-      chrome.tabs.get.mockResolvedValue({ url: 'chrome-error://chromewebdata' });
-
-      await service.migrateLegacyHighlights(1, 'https://example.com', 'key');
-
-      expect(mockInjectionService.injectWithResponse).not.toHaveBeenCalled();
-    });
-
-    it('should save migrated data to storage', async () => {
-      mockInjectionService.injectWithResponse.mockResolvedValue({
-        migrated: true,
-        data: [{ id: '1', text: 'highlight' }],
-        foundKey: 'highlights_old',
-      });
-      chrome.storage.local.set.mockImplementation(_data => Promise.resolve());
-
-      await service.migrateLegacyHighlights(
-        1,
-        'https://example.com',
-        'highlights_https://example.com'
-      );
-
-      expect(mockInjectionService.injectWithResponse).toHaveBeenCalled();
-      expect(chrome.storage.local.set).toHaveBeenCalledWith({
-        'highlights_https://example.com': [{ id: '1', text: 'highlight' }],
-      });
-      expect(mockInjectionService.injectHighlightRestore).toHaveBeenCalledWith(1);
-    });
-
-    it('should handle recoverable errors gracefully', async () => {
-      mockInjectionService.injectWithResponse.mockRejectedValue(
-        new Error('Cannot access contents of page')
-      );
-
-      await service.migrateLegacyHighlights(
-        1,
-        'https://example.com',
-        'highlights_https://example.com'
-      );
-
-      expect(mockLogger.warn).toHaveBeenCalled();
-      expect(mockLogger.warn.mock.calls[0][0]).toMatch(
-        /Migration skipped due to recoverable error:?/
-      );
-    });
-
-    it('should log non-recoverable errors', async () => {
-      mockInjectionService.injectWithResponse.mockRejectedValue(new Error('Unknown error'));
-
-      await service.migrateLegacyHighlights(
-        1,
-        'https://example.com',
-        'highlights_https://example.com'
-      );
-
-      expect(mockLogger.error).toHaveBeenCalled();
-    });
-  });
-
-  describe('waitForTabCompilation', () => {
-    // Helper to flush potential microtasks.
-    // We need to wait for the async _waitForTabCompilation method to proceed past its
-    // initial `await chrome.tabs.get()` and register the event listeners.
-    // A single await might not be enough depending on the internal promise chain.
-    const flushMicrotasks = async () => {
-      await Promise.resolve();
-      await Promise.resolve();
-    };
-
-    it('應該在標籤頁已完成載入時直接返回', async () => {
-      chrome.tabs.get.mockResolvedValue({ id: 1, status: 'complete' });
-      const res = await service._waitForTabCompilation(1);
-      expect(res.status).toBe('complete');
-    });
-
-    it('應該在已有待處理監聽器時返回 null', async () => {
-      service.pendingListeners.set(1, {});
-      chrome.tabs.get.mockResolvedValue({ id: 1, status: 'loading' });
-      const res = await service._waitForTabCompilation(1);
-      expect(res).toBeNull();
-    });
-
-    it('應該處理標籤頁更新事件', async () => {
-      chrome.tabs.get.mockResolvedValue({ id: 1, status: 'loading' });
-
-      let updateCallback;
-      chrome.tabs.onUpdated.addListener.mockImplementation(cb => {
-        updateCallback = cb;
-      });
-
-      const promise = service._waitForTabCompilation(1);
-
-      // 等待非同步操作推進到監聽器註冊階段
-      await flushMicrotasks();
-
-      // 觸發更新
-      if (typeof updateCallback === 'function') {
-        updateCallback(1, { status: 'complete' });
-      }
-
-      const res = await promise;
-      expect(res.status).toBe('complete');
-    });
-
-    it('應該處理標籤頁移除事件', async () => {
-      chrome.tabs.get.mockResolvedValue({ id: 1, status: 'loading' });
-
-      let removeCallback;
-      chrome.tabs.onRemoved.addListener.mockImplementation(cb => {
-        removeCallback = cb;
-      });
-
-      const promise = service._waitForTabCompilation(1);
-
-      // 等待非同步操作推進到監聽器註冊階段
-      await flushMicrotasks();
-
-      // 觸發移除
-      if (typeof removeCallback === 'function') {
-        removeCallback(1);
-      }
-
-      const res = await promise;
-      expect(res).toBeNull();
-    });
-
-    it('應該在超時後返回 null', async () => {
-      jest.useFakeTimers();
-      try {
-        chrome.tabs.get.mockResolvedValue({ id: 1, status: 'loading' });
-
-        const promise = service._waitForTabCompilation(1);
-
-        // 等待非同步操作推進到內部 Promise 建立
-        await flushMicrotasks();
-
-        jest.advanceTimersByTime(11_000); // 大於 10s
-
-        const res = await promise;
-        expect(res).toBeNull();
-        expect(mockLogger.warn).toHaveBeenCalledWith(expect.stringContaining('timeout'));
-      } finally {
-        jest.useRealTimers();
-      }
-    });
-  });
-
-  describe('Concurrency & Edge Cases', () => {
-    it('updateTabStatus 應該防止並發處理同一個標籤頁 (Line 62)', async () => {
-      service.processingTabs.set(1, Date.now());
-      await service.updateTabStatus(1, 'https://example.com');
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        expect.stringContaining('already being processed')
-      );
-    });
-
-    it('migrateLegacyHighlights 應該處理腳本回報的錯誤 (Line 277)', async () => {
-      chrome.tabs.get.mockResolvedValue({ id: 1, url: 'https://example.com' });
-      mockInjectionService.injectWithResponse.mockResolvedValue({ error: 'Injected script fail' });
-
-      await service.migrateLegacyHighlights(1, 'https://example.com', 'key');
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        expect.stringContaining('Migration script reported error'),
-        expect.anything()
-      );
-    });
-  });
-
-  describe('Wrappers (waitForTabComplete, queryTabs, createTab, removeTab)', () => {
-    it('queryTabs should wrap chrome.tabs.query', async () => {
-      chrome.tabs.query = jest.fn().mockResolvedValue([]);
-      await service.queryTabs({ active: true });
-      expect(chrome.tabs.query).toHaveBeenCalledWith({ active: true });
-    });
-
-    it('createTab should wrap chrome.tabs.create', async () => {
-      chrome.tabs.create = jest.fn().mockResolvedValue({});
-      await service.createTab({ url: 'https://test.com' });
-      expect(chrome.tabs.create).toHaveBeenCalledWith({ url: 'https://test.com' });
-    });
-
-    it('removeTab should wrap chrome.tabs.remove', async () => {
-      chrome.tabs.remove = jest.fn().mockResolvedValue();
-      await service.removeTab(1);
-      expect(chrome.tabs.remove).toHaveBeenCalledWith(1);
-    });
-
-    it('waitForTabComplete should call _waitForTabCompilation', async () => {
-      const waitForSpy = jest.spyOn(service, '_waitForTabCompilation').mockResolvedValue(true);
-      await service.waitForTabComplete(1);
-      expect(waitForSpy).toHaveBeenCalledWith(1);
-    });
-  });
-
-  describe('resolveTabUrl edges', () => {
-    afterEach(() => {
-      urlUtils.resolveStorageUrl.mockRestore?.();
-      urlUtils.isRootUrl.mockRestore?.();
-    });
-
-    it('應處理 isRootUrl 為 true 的情況', async () => {
-      urlUtils.resolveStorageUrl.mockReturnValueOnce('https://example.com/');
-      urlUtils.isRootUrl.mockReturnValueOnce(true);
-
-      const res = await service.resolveTabUrl(1, 'https://example.com/?some=param');
-
-      expect(res.hasStableUrl).toBe(false);
-      expect(res.stableUrl).toBe('https://example.com/?some=param'); // fallbacks to originalUrl
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        expect.stringContaining('Blocked root URL as stableUrl'),
-        expect.anything()
-      );
-    });
-
-    it('應執行 migrationService.migrateStorageKey', async () => {
-      urlUtils.resolveStorageUrl.mockReturnValueOnce('https://example.com/stable');
-      urlUtils.isRootUrl.mockReturnValueOnce(false);
-
-      const mockMigrationService = { migrateStorageKey: jest.fn().mockResolvedValue(true) };
-      const res = await service.resolveTabUrl(1, 'https://example.com/?a=1', mockMigrationService);
-
-      expect(res.hasStableUrl).toBe(true);
-      expect(res.migrated).toBe(true);
-      expect(mockMigrationService.migrateStorageKey).toHaveBeenCalledWith(
-        'https://example.com/stable',
-        'https://example.com/?a=1'
-      );
-    });
-  });
-
-  describe('_sendStableUrl behavior', () => {
-    it('_sendStableUrl 應阻擋 root url 寫入', () => {
-      urlUtils.isRootUrl.mockReturnValueOnce(true);
-
-      service._sendStableUrl(1, 'https://example.com/');
-      expect(mockLogger.warn).toHaveBeenCalledWith(
-        expect.stringContaining('Blocked SET_STABLE_URL'),
-        expect.anything()
-      );
-      expect(chrome.tabs.sendMessage).not.toHaveBeenCalled();
-    });
-
-    it('_sendStableUrl 應正常發送訊息', () => {
-      urlUtils.isRootUrl.mockReturnValueOnce(false);
-      chrome.tabs.sendMessage.mockReturnValue(Promise.resolve());
-
-      service._sendStableUrl(1, 'https://example.com/page');
-      expect(chrome.tabs.sendMessage).toHaveBeenCalledWith(1, {
-        action: 'SET_STABLE_URL',
-        stableUrl: 'https://example.com/page',
-      });
-    });
-  });
-
-  describe('Coverage Improvements', () => {
-    it('_waitForTabCompilation should return null if chrome.tabs.get fails', async () => {
-      chrome.tabs.get.mockRejectedValue(new Error('Get tab failed'));
-      const res = await service._waitForTabCompilation(999);
-      expect(res).toBeNull();
-    });
-
-    it('_waitForTabCompilation should return null if tab is discarded', async () => {
-      chrome.tabs.get.mockResolvedValue({ id: 999, discarded: true });
-      const res = await service._waitForTabCompilation(999);
-      expect(res).toBeNull();
-      expect(mockLogger.debug).toHaveBeenCalledWith(expect.stringContaining('discarded'));
-    });
-
-    it('updateTabStatus should log specific message for "No tab with id" error', async () => {
-      service._verifyAndUpdateStatus = jest
-        .fn()
-        .mockRejectedValue(new Error('No tab with id: 999'));
-      await service.updateTabStatus(999, 'https://example.com');
-      expect(mockLogger.debug).toHaveBeenCalledWith(expect.stringContaining('Tab closed/missing'));
-    });
-
-    it('updateTabStatus should log specific message for "The tab was closed" error', async () => {
-      service._verifyAndUpdateStatus = jest
-        .fn()
-        .mockRejectedValue(new Error('The tab was closed.'));
-      await service.updateTabStatus(999, 'https://example.com');
-      expect(mockLogger.debug).toHaveBeenCalledWith(expect.stringContaining('Tab closed/missing'));
-    });
-  });
-  describe('Stable URL Fallback Logic', () => {
-    test('應正確執行回退查找 (Stable URL Miss -> Original URL Hit)', async () => {
-      /**
-       * 此測試旨在驗證 _updateTabStatusInternal 的「雙查/回退」邏輯：
-       * 當為頁面計算出穩定 URL (Stable URL) 時，應優先查詢該 URL 的標註；
-       * 若穩定 URL 下無數據，應回退到原始 URL (Original URL) 查詢，以確保向後兼容。
-       */
-      const mockTabId = 999;
-      const mockRawUrl = 'https://example.com/?slug=test';
-      const mockStableUrl = 'https://example.com/stable';
-      const mockOriginalUrl = 'https://example.com/';
-
-      // 1. Mock 外部工具函數以模擬 Phase 1 為該頁面生成了不同的穩定 URL
-      // 使用 import 的 mock 對象，避免 require 與手動還原
-      // 重要：在測試結束後還原 mock，避免影響後續測試 (如 getPreloaderData edge cases)
-      const originalImpl = urlUtils.resolveStorageUrl.getMockImplementation();
-      urlUtils.resolveStorageUrl.mockReturnValue(mockStableUrl);
-
-      try {
-        // 2. 配置 service 的 URL 標準化行為
-        service.normalizeUrl = jest.fn().mockReturnValue(mockOriginalUrl);
-
-        // 3. Mock 外部儲存 API：穩定 URL 為空，原始 URL 的 highlights_* key 有數據
-        chrome.storage.local.get.mockImplementation(async keys => {
-          // 穩定 URL 的新舊格式都沒有數據
-          if (
-            keys.includes(`highlights_${mockStableUrl}`) ||
-            keys.includes(`page_${mockStableUrl}`)
-          ) {
-            return {};
-          }
-          // 原始 URL 的舊格式有數據
-          if (keys.includes(`highlights_${mockOriginalUrl}`)) {
-            return { [`highlights_${mockOriginalUrl}`]: [{ text: 'fallback-highlight' }] };
-          }
-          return {};
-        });
-
-        // 4. Mock 其他無關此測試邏輯的內部步驟，以隔離並專注於回退邏輯檢測
-        service.getPreloaderData = jest.fn().mockResolvedValue(null);
-        service._verifyAndUpdateStatus = jest.fn().mockResolvedValue();
-        service._waitForTabCompilation = jest.fn().mockResolvedValue({ id: mockTabId });
-        service.injectionService = { ensureBundleInjected: jest.fn().mockResolvedValue() };
-
-        // 5. 執行測試
-        await service._updateTabStatusInternal(mockTabId, mockRawUrl);
-
-        // 6. 驗證：Phase 4 (tighten) 在 _persistUrlAliasIfNeeded 加入 evidence check 為第 1 個 get；
-        //         resolver 隨後對 stable / original 各做一次 get（第 2、3 個）。
-        expect(chrome.storage.local.get).toHaveBeenNthCalledWith(
-          1,
-          expect.arrayContaining([`page_${mockStableUrl}`, `highlights_${mockStableUrl}`])
-        );
-        // 第 2 個 get：resolver 優先查 stable URL 的所有 keys
-        expect(chrome.storage.local.get).toHaveBeenNthCalledWith(
-          2,
-          expect.arrayContaining([
-            `${URL_ALIAS_PREFIX}${mockStableUrl}`,
-            `page_${mockStableUrl}`,
-            `highlights_${mockStableUrl}`,
-          ])
-        );
-        // 第 3 個 get：stable miss 後回退到 original URL 的所有 keys
-        expect(chrome.storage.local.get).toHaveBeenNthCalledWith(
-          3,
-          expect.arrayContaining([
-            `${URL_ALIAS_PREFIX}${mockOriginalUrl}`,
-            `page_${mockOriginalUrl}`,
-            `highlights_${mockOriginalUrl}`,
-          ])
-        );
-
-        // 驗證最終成功觸發了注入
-        expect(service.injectionService.ensureBundleInjected).toHaveBeenCalledWith(mockTabId);
-      } finally {
-        // 還原 Mock
-        if (originalImpl) {
-          urlUtils.resolveStorageUrl.mockImplementation(originalImpl);
-        } else {
-          urlUtils.resolveStorageUrl.mockReset(); // 或者 mockBack to default implementation if needed
-          urlUtils.resolveStorageUrl.mockImplementation(url => url); // Restore default mock behavior defined at top of file
-        }
-      }
-    });
-  });
-
-  describe('getPreloaderData', () => {
-    it('應該成功獲取 Preloader 數據', async () => {
-      const mockData = { shortlink: 'https://example.com/p=1', nextRouteInfo: null };
-      chrome.tabs.sendMessage.mockImplementation((_tabId, _msg, cb) => cb(mockData));
-
-      const result = await service.getPreloaderData(1);
-      expect(result).toEqual(mockData);
-      expect(chrome.tabs.sendMessage).toHaveBeenCalledWith(
-        1,
-        { action: 'PING' },
-        expect.any(Function)
-      );
-    });
-
-    it('應該在超時後返回 null', async () => {
-      jest.useFakeTimers();
-      try {
-        chrome.tabs.sendMessage.mockImplementation(() => {
-          // 不調用 callback，模擬無響應
-        });
-
-        const promise = service.getPreloaderData(1);
-        jest.advanceTimersByTime(1000); // 超過 500ms
-        const result = await promise;
-
-        expect(result).toBeNull();
-      } finally {
-        jest.useRealTimers();
-      }
-    });
-
-    it('應該在 runtime.lastError 時返回 null', async () => {
-      chrome.tabs.sendMessage.mockImplementation((_tabId, _msg, cb) => {
-        chrome.runtime.lastError = { message: 'Port closed' };
-        cb();
-      });
-
-      const result = await service.getPreloaderData(1);
-      expect(result).toBeNull();
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        expect.stringContaining('Failed to get preloader data')
-      );
-
-      // 測試結束後清理，避免污染後續測試
-      delete chrome.runtime.lastError;
-    });
-
-    it('應該在 sendMessage 拋出異常時返回 null', async () => {
-      chrome.tabs.sendMessage.mockImplementation(() => {
-        throw new Error('API Error');
-      });
-
-      const result = await service.getPreloaderData(1);
-      expect(result).toBeNull();
-      expect(mockLogger.debug).toHaveBeenCalledWith(
-        expect.stringContaining('Failed to get preloader data')
-      );
-    });
-
-    describe('Coverage Improvements (Edge Cases)', () => {
-      test('應正確處理 getPreloaderData 返回 null 的情況 (回退到 normalizeUrl)', async () => {
-        service.getPreloaderData = jest.fn().mockResolvedValue(null);
-        // Ensure normalizeUrl uses default behavior or specific mock
-        service.normalizeUrl = jest.fn(url => url);
-
-        const url = 'https://example.com/test';
-        // resolveStorageUrl util will return normalizeUrl(url) if preloaderData is null
-        // We need to ensure resolveStorageUrl is not mocked to return something else from previous tests OR restore it
-        // The previous test mocked resolveStorageUrl, so we must ensure it's restored.
-        // The previous test does restore it in 'finally' logical block (lines 635-637 of original file)
-
-        const result = await service.resolveTabUrl(999, url);
-
-        expect(result.stableUrl).toBe(url); // resolveStorageUrl fallback
-        expect(result.originalUrl).toBe(url);
-        expect(result.hasStableUrl).toBe(false);
-      });
-
-      test('應正確執行 _verifyAndUpdateStatus 的回退查詢邏輯 (Stable URL Miss -> Original URL Hit)', async () => {
-        const tabId = 999;
-        const normUrl = 'https://example.com/stable';
-        const fallbackUrl = 'https://example.com/';
-
-        // Setup mocks
-        service.getSavedPageData = jest
-          .fn()
-          .mockResolvedValueOnce(null) // First call for normUrl returns null
-          .mockResolvedValueOnce({ notionPageId: 'page-123', lastVerifiedAt: Date.now() }); // Second call for fallbackUrl returns data
-
-        service._updateBadgeStatus = jest.fn().mockResolvedValue();
-
-        // Execute private method
-        await service._verifyAndUpdateStatus(tabId, normUrl, fallbackUrl);
-
-        // Verify
-        expect(service.getSavedPageData).toHaveBeenNthCalledWith(1, normUrl);
-        expect(service.getSavedPageData).toHaveBeenNthCalledWith(2, fallbackUrl);
-        expect(service._updateBadgeStatus).toHaveBeenCalledWith(
-          tabId,
-          expect.objectContaining({ notionPageId: 'page-123' })
-        );
-      });
-    });
-  });
-
-  describe('LIFECYCLE & LISTENERS', () => {
-    it('應該處理 chrome.tabs.onActivated 事件', async () => {
-      chrome.tabs.get.mockResolvedValue({ id: 10, url: 'https://example.com/activated' });
-
-      service.setupListeners();
-      const activatedCallback = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-
-      await activatedCallback({ tabId: 10, windowId: 1 });
-
-      // 驗證 updateTabStatus 被呼叫
-      expect(chrome.tabs.get).toHaveBeenCalledWith(10);
-    });
-
-    it('應該在 chrome.tabs.onActivated 失敗時靜默處理', async () => {
-      chrome.tabs.get.mockRejectedValue(new Error('Tab not found'));
-
-      service.setupListeners();
-      const activatedCallback = chrome.tabs.onActivated.addListener.mock.calls[0][0];
-
-      await expect(activatedCallback({ tabId: 999, windowId: 1 })).resolves.toBeUndefined();
-      expect(mockLogger.debug).toHaveBeenCalled();
-    });
-
-    it('應該處理 chrome.tabs.onUpdated 事件', () => {
-      jest.useFakeTimers();
-
-      // Spy on the public method we expect to be called
-      const updateStatusSpy = jest.spyOn(service, 'updateTabStatus').mockImplementation(() => {});
-
-      service.setupListeners();
-      const updatedCallback = chrome.tabs.onUpdated.addListener.mock.calls[0][0];
-
-      updatedCallback(1, { status: 'complete' }, { id: 1, url: 'https://example.com/updated' });
-
-      // 推進時間：等待 STATUS_UPDATE_DELAY_MS (100ms) 觸發延遲更新
-      jest.advanceTimersByTime(200);
-
-      // 驗證延遲後的行為：應該呼叫 updateTabStatus
-      expect(updateStatusSpy).toHaveBeenCalledWith(1, 'https://example.com/updated');
-
-      jest.useRealTimers();
-      updateStatusSpy.mockRestore();
-    });
-  });
-
-  describe('_migrationScript isolated', () => {
-    let originalHref;
-    let store = {};
-
-    beforeEach(() => {
-      originalHref = globalThis.location.href;
-      // Use history.pushState instead of deleting location
-      globalThis.history.pushState({}, 'Test Title', 'https://example.com/test?utm_source=123');
-
-      store = {};
-      jest.spyOn(Storage.prototype, 'getItem').mockImplementation(k => store[k] || null);
-      jest.spyOn(Storage.prototype, 'setItem').mockImplementation((k, v) => (store[k] = String(v)));
-      jest.spyOn(Storage.prototype, 'removeItem').mockImplementation(k => delete store[k]);
-      jest.spyOn(Storage.prototype, 'key').mockImplementation(i => Object.keys(store)[i] || null);
-      jest
-        .spyOn(Storage.prototype, 'length', 'get')
-        .mockImplementation(() => Object.keys(store).length);
-    });
-
-    afterEach(() => {
-      globalThis.history.pushState({}, 'Original', originalHref);
-      jest.restoreAllMocks();
-    });
-
-    it('應處理 trailing slash', () => {
-      globalThis.history.pushState({}, 'Slash', 'https://example.com/test/');
-      store['highlights_https://example.com/test'] = '[{"text":"hi"}]';
-      const res = _migrationScript([]);
-      expect(res.migrated).toBe(true);
-    });
-
-    it('應處理 normalize 拋出例外', () => {
-      globalThis.history.pushState({}, 'Test', 'https://example.com/test');
-      const originalURL = globalThis.URL;
-      try {
-        globalThis.URL = jest.fn().mockImplementation(() => {
-          throw new Error('mock error');
-        });
-
-        store['highlights_https://example.com/test'] = '[{"text":"hi"}]';
-
-        const res = _migrationScript([]);
-        expect(res.migrated).toBe(true);
-      } finally {
-        globalThis.URL = originalURL;
-      }
-    });
-
-    it('應處理取得的 raw 資料為 falsy 的情況', () => {
-      globalThis.history.pushState({}, 'Test', 'https://example.com/test');
-      globalThis.localStorage.setItem('highlights_https://example.com/test', ''); // empty string
-      const res = _migrationScript([]);
-      expect(res.migrated).toBe(false);
-    });
-
-    it('應找不到 key 時返回 migrated: false', () => {
-      const res = _migrationScript(['utm_source']);
-      expect(res.migrated).toBe(false);
-    });
-
-    it('應成功遷移資料 (Highlights_ 開頭)', () => {
-      globalThis.localStorage.setItem(
-        'highlights_https://example.com/test',
-        JSON.stringify([{ text: 'hi' }])
-      );
-      const res = _migrationScript(['utm_source']);
-      expect(res.migrated).toBe(true);
-      expect(res.data[0].text).toBe('hi');
-      expect(globalThis.localStorage.getItem('highlights_https://example.com/test')).toBeNull(); // removed
-    });
-
-    it('應成功遷移資料 (Fallback 遍歷)', () => {
-      globalThis.history.pushState({}, 'Other', 'https://example.com/other');
-      globalThis.localStorage.setItem('highlights_some-old-key', JSON.stringify([{ text: 'hi2' }]));
-      const res = _migrationScript([]);
-      expect(res.migrated).toBe(true);
-      expect(res.data[0].text).toBe('hi2');
-    });
-
-    it('解析錯誤應被捕捉並返回 false', () => {
-      globalThis.localStorage.setItem('highlights_https://example.com/test', 'invalid-json');
-      const consoleSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
-      const res = _migrationScript(['utm_source']);
-      expect(res.migrated).toBe(false);
-      expect(consoleSpy).toHaveBeenCalled();
-      consoleSpy.mockRestore();
-    });
-
-    it('異常應被捕獲並返回 false', () => {
-      Storage.prototype.getItem.mockImplementation(() => {
-        throw new Error('simulate error');
-      });
-      globalThis.localStorage.setItem('highlights_https://example.com/test', 'some json');
-      const consoleSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
-      const res = _migrationScript(['utm_source']);
-      expect(res.migrated).toBe(false);
-      expect(res.error).toBeDefined();
-      consoleSpy.mockRestore();
     });
   });
 });
